+++ src/main/java/com/timelink/time_link/dto/Student/StudentResponseDTO.java
@@ -3,10 +3,10 @@ package com.timelink.time_link.dto.Student;
 import java.time.LocalDate;
 
 public record StudentResponseDTO(
-        Long id,
+        Integer id,
         String name,
         Boolean active,
         LocalDate dateBirth,
         String username,
-        Long groupId
+        Integer groupId
 ) {}
